requirements.txt
streamlit>=1.36
openai>=1.40.0
python-dotenv>=1.0.1
pydantic>=2.7
rich>=13.7

# LangChain + RAG
langchain>=0.2.12
langchain-openai>=0.1.20
langchain-text-splitters>=0.2.2
langchain-community>=0.2.11
chromadb>=0.5.5

# Parsing
pypdf>=4.3.1
python-docx>=1.1.2
